chore(refactor): create dedicated file for errors/types (refs #47)
